555 split edge pairing into two files
