+++ rubikscubennnsolver/RubiksCubeNNNEvenEdges.py
@@ -1,7 +1,8 @@
 
 from rubikscubennnsolver import RubiksCube, ImplementThis
 from rubikscubennnsolver.RubiksCube444 import RubiksCube444, solved_444
-from rubikscubennnsolver.RubiksCube555 import RubiksCube555, solved_555
+from rubikscubennnsolver.RubiksCube555 import solved_555
+from rubikscubennnsolver.RubiksCube555ForNNN import RubiksCube555ForNNN
 import logging
 import sys
 
@@ -22,7 +23,7 @@ class RubiksCubeNNNEvenEdges(RubiksCube):
 
     def get_fake_555(self):
         if self.fake_555 is None:
-            self.fake_555 = RubiksCube555(solved_555, 'URFDLB')
+            self.fake_555 = RubiksCube555ForNNN(solved_555, 'URFDLB')
             self.fake_555.lt_init()
             self.fake_555.enable_print_cube = False
         else:
@@ -205,7 +206,7 @@ class RubiksCubeNNNEvenEdges(RubiksCube):
         else:
             # For 8x8x8 and up the centers will already be solved so use the
             # older but faster L4E edge pairing code.
-            fake_555.reduce_333_via_l4e(fake_555=True)
+            fake_555.reduce_333(fake_555=True)
 
         fake_555.print_cube()
